client/requirements.txt
PyQt6>=6.6.0
mss>=9.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
httpx>=0.27.0
qasync>=0.27.0
